update by admin 2/5

+++ ums_backend/src/main/java/com/university/service/admin/NganhAdminService.java
@@ -65,6 +65,10 @@ public class NganhAdminService {
     }
 
     public void delete(UUID id) {
+        Nganh nganh = nganhRepository.findById(id).orElseThrow();
+        if (nganh == null) {
+            throw new EntityNotFoundException("Nganh ko ton tai");
+        }
         nganhRepository.deleteById(id);
     }
 }
